+++ InfiniteGraphTraversalAssignment/TraversalStrategy.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using System.Numerics;
 
 // TODO: event clean-up
 
@@ -10,13 +11,13 @@ namespace InfiniteGraphTraversalAssignment
     {
         protected Graph graph = null;
 
-        public delegate void AllChildrenVisitedHandler(int nodeIndex);
+        public delegate void AllChildrenVisitedHandler(BigInteger nodeIndex);
         public virtual event AllChildrenVisitedHandler OnAllChildrenVisited;
 
-        public delegate void HasNoChildrenHandler(int nodeIndex);
+        public delegate void HasNoChildrenHandler(BigInteger nodeIndex);
         public virtual event HasNoChildrenHandler OnHasNoChildren;
 
         public abstract bool IsFinished { get; protected set; }
-        public abstract int Execute();
+        public abstract BigInteger Execute();
     }
 }
